MAINT:verifylib, toto-verify: libify in_toto_verify
Integrators of Toto may want to directly access a "seamless"
verification routine to avoid any misunderstandings on errors upon
implementation. Move "holistic" in_toto_verify function to verifylib so
integrators can use it.
