+++ toto/verifylib.py
@@ -39,6 +39,7 @@ import toto.models.link
 import toto.ssl_crypto.keys
 from toto.exceptions import RuleVerficationFailed
 from toto.models.common import ComparableHashDict
+from toto.models.layout import Layout
 import toto.log as log
 
 
@@ -585,6 +586,78 @@ def verify_all_item_rules(items, links, target_links=None):
         link.materials, target_links)
     verify_item_rules(item.name, item.product_matchrules,
         link.products, target_links)
+
+def in_toto_verify(layout_path, layout_key_paths):
+  """Loads layout file and layout keys from disk and performs all in-toto
+  verifications."""
+
+  try:
+    log.doing("load layout...")
+    layout = Layout.read_from_file(layout_path)
+  except Exception, e:
+    _die("in load layout - %s" % e)
+
+  try:
+    log.doing("verify layout expiration")
+    verify_layout_expiration(layout)
+  except Exception, e:
+    _die("in verify layout expiration - %s" % e)
+
+  try:
+    log.doing("load layout keys...")
+    layout_key_dict = toto.util.import_rsa_public_keys_from_files_as_dict(
+        layout_key_paths)
+  except Exception, e:
+    _die("in load layout keys - %s" % e)
+
+  try:
+    log.doing("verify layout signatures...")
+    verify_layout_signatures(layout, layout_key_dict)
+  except Exception, e:
+    _die("in verify layout signatures - %s" % e)
+
+  try:
+    log.doing("load link metadata for steps...")
+    step_link_dict = layout.import_step_metadata_from_files_as_dict()
+  except Exception, e:
+    _die("in load link metadata - %s" % e)
+
+  try:
+    log.doing("verify all step command alignments...")
+    verify_all_steps_command_alignment(layout, step_link_dict)
+  except Exception, e:
+    _die("command alignments - %s" % e)
+
+  try:
+    log.doing("verify signatures for all links...")
+    verify_all_steps_signatures(layout, step_link_dict)
+  except Exception, e:
+    _die("in verify link signatures - %s" % e)
+
+  try:
+    log.doing("run all inspections...")
+    inspection_link_dict = run_all_inspections(layout)
+  except Exception, e:
+    _die("in run inspections - %s" % e)
+
+  try:
+    log.doing("verify all step matchrules...")
+    verify_all_item_rules(layout.steps, step_link_dict)
+  except Exception, e:
+    _die("in verify all step matchrules - %s" % e)
+
+  try:
+    log.doing("verify all inspection matchrules...")
+    verify_all_item_rules(layout.inspect, inspection_link_dict,
+        step_link_dict)
+  except Exception, e:
+    _die("in verify all inspection matchrules - %s" % e)
+
+  log.passing("all verification")
+
+def _die(msg, exitcode=1):
+  log.failing(msg)
+  sys.exit(exitcode)
